Add files via upload
The Application of Fuzzy Network Graph in Hospital for Disease Prediction using Symptoms represents a groundbreaking approach to healthcare diagnostics by combining the power of fuzzy logic, network graphs, and symptom-based prediction. This project is designed to enhance the accuracy and efficiency of disease prediction within a hospital setting, ultimately aiding healthcare professionals in making timely and informed decisions.